Move lexing code into module
